+++ compiler/program.py
@@ -1,8 +1,5 @@
-import argparse
 import subprocess
 from typing import List
-from compiler.asm import initialize_asm, generate_asm
-from compiler.compile import compile_asm, link_object_file
 from compiler.defs import TokenType, Token, OpType, Program, Op, Intrinsic
 
 def intrinsic_exists(token: str) -> bool:
@@ -44,20 +41,6 @@ def generate_program(tokens = List[Token]) -> Program:
         operand: Op = Op(id, op_type, token)
         program.append(operand)
     return program
-
-def compile_code(tokens: List[Token], input_file: str, output_file: str) -> None:
-    asm_file: str = input_file.replace('.torth', '.asm')
-    program: Program = generate_program(tokens)
-    initialize_asm(asm_file)
-    generate_asm(program, asm_file)
-    compile_asm(asm_file)
-    link_object_file(asm_file.replace('.asm', '.o'), output_file)
-
-def remove_compilation_files(input_file: str, args: argparse.Namespace) -> None:
-    input_file_extensionless: str = input_file.split('.')[0]
-    subprocess.run(['rm', '-f', f'{input_file_extensionless}.o'])
-    if not args.save_asm:
-        subprocess.run(['rm', '-f', f'{input_file_extensionless}.asm'])
 
 def run_code(exe_file: str) -> None:
     subprocess.run([f'./{exe_file}'])
